+++ internal/machine/spectrum/bus_test.go
@@ -1,0 +1,11 @@
+// Package spectrum implements the ZX Spectrum 48K machine logic.
+package spectrum
+
+import (
+	"mcs/testutils/dsl"
+	"testing"
+)
+
+func TestBus(t *testing.T) {
+	dsl.RunScenarios(t, busScenarios)
+}